# Improve allocation of properties