--- a/builder.go
+++ b/builder.go
@@ -18,7 +18,7 @@ type ErrorBuilder struct {
 }
 
 // NewErrorBuilder creates error builder from an existing error type.
-func NewErrorBuilder(t *Type) *ErrorBuilder {
+func NewErrorBuilder(t *Type) ErrorBuilder {
 	getMode := func() callStackBuildMode {
 		if !t.modifiers.CollectStackTrace() {
 			return stackTraceOmit
@@ -26,7 +26,7 @@ func NewErrorBuilder(t *Type) *ErrorBuilder {
 		return stackTraceCollect
 	}
 
-	return &ErrorBuilder{
+	return ErrorBuilder{
 		errorType:     t,
 		mode:          getMode(),
 		isTransparent: t.modifiers.Transparent(),
@@ -37,7 +37,7 @@ func NewErrorBuilder(t *Type) *ErrorBuilder {
 // For non-errorx errors, a stack trace is collected.
 // Otherwise, it is inherited by default, as error wrapping is typically performed 'en passe'.
 // Note that even if an original error explicitly omitted the stack trace, it could be added on wrap.
-func (eb *ErrorBuilder) WithCause(err error) *ErrorBuilder {
+func (eb ErrorBuilder) WithCause(err error) ErrorBuilder {
 	eb.cause = err
 	if Cast(err) != nil {
 		eb.mode = stackTraceBorrow
@@ -50,7 +50,7 @@ func (eb *ErrorBuilder) WithCause(err error) *ErrorBuilder {
 // Transparent wrap hides the current error type from the type checks and exposes the error type of the cause instead.
 // The same holds true for traits, and the dynamic properties are visible from both cause and transparent wrapper.
 // Note that if the cause error is non-errorx, transparency will still hold, type check against wrapper will still fail.
-func (eb *ErrorBuilder) Transparent() *ErrorBuilder {
+func (eb ErrorBuilder) Transparent() ErrorBuilder {
 	if eb.cause == nil {
 		panic("wrong builder usage: wrap modifier without non-nil cause")
 	}
@@ -64,7 +64,7 @@ func (eb *ErrorBuilder) Transparent() *ErrorBuilder {
 // This is typically a way to handle an error received from another goroutine - say, a worker pool.
 // When stack traces overlap, formatting makes a conservative attempt not to repeat itself,
 // preserving the *original* stack trace in its entirety.
-func (eb *ErrorBuilder) EnhanceStackTrace() *ErrorBuilder {
+func (eb ErrorBuilder) EnhanceStackTrace() ErrorBuilder {
 	if eb.cause == nil {
 		panic("wrong builder usage: wrap modifier without non-nil cause")
 	}
@@ -81,7 +81,7 @@ func (eb *ErrorBuilder) EnhanceStackTrace() *ErrorBuilder {
 // WithConditionallyFormattedMessage provides a message for an error in flexible format, to simplify its usages.
 // Without args, leaves the original message intact, so a message may be generated or provided externally.
 // With args, a formatting is performed, and it is therefore expected a format string to be constant.
-func (eb *ErrorBuilder) WithConditionallyFormattedMessage(message string, args ...interface{}) *ErrorBuilder {
+func (eb ErrorBuilder) WithConditionallyFormattedMessage(message string, args ...interface{}) ErrorBuilder {
 	if len(args) == 0 {
 		eb.message = message
 	} else {
@@ -92,7 +92,7 @@ func (eb *ErrorBuilder) WithConditionallyFormattedMessage(message string, args .
 }
 
 // Create returns an error with specified params.
-func (eb *ErrorBuilder) Create() *Error {
+func (eb ErrorBuilder) Create() *Error {
 	return &Error{
 		errorType:   eb.errorType,
 		message:     eb.message,
@@ -112,7 +112,7 @@ const (
 	stackTraceOmit    callStackBuildMode = 4
 )
 
-func (eb *ErrorBuilder) assembleStackTrace() *stackTrace {
+func (eb ErrorBuilder) assembleStackTrace() *stackTrace {
 	switch eb.mode {
 	case stackTraceCollect:
 		return eb.collectOriginalStackTrace()
@@ -127,11 +127,11 @@ func (eb *ErrorBuilder) assembleStackTrace() *stackTrace {
 	}
 }
 
-func (eb *ErrorBuilder) collectOriginalStackTrace() *stackTrace {
+func (eb ErrorBuilder) collectOriginalStackTrace() *stackTrace {
 	return collectStackTrace()
 }
 
-func (eb *ErrorBuilder) borrowStackTraceFromCause() *stackTrace {
+func (eb ErrorBuilder) borrowStackTraceFromCause() *stackTrace {
 	originalStackTrace := eb.extractStackTraceFromCause(eb.cause)
 	if originalStackTrace != nil {
 		return originalStackTrace
@@ -139,7 +139,7 @@ func (eb *ErrorBuilder) borrowStackTraceFromCause() *stackTrace {
 	return collectStackTrace()
 }
 
-func (eb *ErrorBuilder) combineStackTraceWithCause() *stackTrace {
+func (eb ErrorBuilder) combineStackTraceWithCause() *stackTrace {
 	currentStackTrace := collectStackTrace()
 
 	originalStackTrace := eb.extractStackTraceFromCause(eb.cause)
@@ -150,7 +150,7 @@ func (eb *ErrorBuilder) combineStackTraceWithCause() *stackTrace {
 	return currentStackTrace
 }
 
-func (eb *ErrorBuilder) extractStackTraceFromCause(cause error) *stackTrace {
+func (eb ErrorBuilder) extractStackTraceFromCause(cause error) *stackTrace {
 	if typedCause := Cast(cause); typedCause != nil {
 		return typedCause.stackTrace
 	}