--- a/id.go
+++ b/id.go
@@ -1,23 +1,13 @@
 package errorx
 
 import (
-	"sync"
+	"sync/atomic"
 )
 
-var internalID uint64 = 0x123456789abcdef0
-var idMtx sync.Mutex
+var internalID uint64
 
 // nextInternalID creates next unique id for errorx entities.
 // All equality comparison should take id into account, lest there be some false positive matches.
 func nextInternalID() uint64 {
-	idMtx.Lock()
-	// xorshift
-	x := internalID
-	x ^= x << 13
-	x ^= x >> 7
-	x ^= x << 17
-	id := x + internalID
-	internalID = x
-	idMtx.Unlock()
-	return id
+	return atomic.AddUint64(&internalID, 1)
 }