--- a/property.go
+++ b/property.go
@@ -8,7 +8,11 @@ import (
 // Property value belongs to an error instance only, never inherited from a type.
 // Property visibility is hindered by Wrap, preserved by Decorate.
 type Property struct {
-	id    int64
+	*property
+}
+
+type property struct {
+	id    uint64
 	label string
 }
 
@@ -70,8 +74,56 @@ var (
 )
 
 func newProperty(label string) Property {
-	return Property{
-		id:    nextInternalID(),
-		label: label,
+	p := Property{
+		&property{
+			id:    nextInternalID(),
+			label: label,
+		},
+	}
+	return p
+}
+
+// propertyMap represents map of properties.
+// Compared to builtin type, it uses less allocations and reallocations on copy.
+// It is simple binary search tree that relies on property id randomness for balancing.
+type propertyMap struct {
+	p     Property
+	value interface{}
+	left  *propertyMap
+	right *propertyMap
+}
+
+// withProperty creates new map with property added or overwritten
+func (pm *propertyMap) with(p Property, value interface{}) *propertyMap {
+	if pm == nil {
+		return &propertyMap{p: p, value: value}
+	} else if pm.p == p {
+		return &propertyMap{
+			p:     p,
+			value: value,
+			left:  pm.left,
+			right: pm.right,
+		}
+	} else {
+		copy := *pm
+		if copy.p.id < p.id {
+			copy.right = copy.right.with(p, value)
+		} else {
+			copy.left = copy.left.with(p, value)
+		}
+		return &copy
+	}
+}
+
+func (pm *propertyMap) get(p Property) (value interface{}, ok bool) {
+	switch {
+	case pm == nil:
+		return nil, false
+	case pm.p == p:
+		return pm.value, true
+	case pm.p.id < p.id:
+		return pm.right.get(p)
+	default:
+		return pm.left.get(p)
 	}
 }