--- a/property_test.go
+++ b/property_test.go
@@ -125,3 +125,34 @@ func BenchmarkAllocProperty(b *testing.B) {
 		})
 	}
 }
+
+var sum int
+
+func BenchmarkGetProperty(b *testing.B) {
+	const N = 9
+	var properties = []Property{}
+	for j := 0; j < N; j++ {
+		n := fmt.Sprintf("props%d", j)
+		properties = append(properties, RegisterProperty(n))
+		b.Run(n, func(b *testing.B) {
+			err := testTypeSilent.New("test")
+			for i := 0; i < j; i++ {
+				err = err.WithProperty(properties[i], 42)
+			}
+			for k := 0; k < b.N; k++ {
+				v, ok := err.Property(testProperty0)
+				if ok {
+					sum += v.(int)
+				}
+				v, ok = err.Property(properties[j])
+				if ok {
+					sum += v.(int)
+				}
+				v, ok = err.Property(properties[0])
+				if ok {
+					sum += v.(int)
+				}
+			}
+		})
+	}
+}