```
benchmark                         old ns/op     new ns/op     delta
BenchmarkAllocProperty/props0     100           83.0          -17.00%
BenchmarkAllocProperty/props1     319           172           -46.08%
BenchmarkAllocProperty/props2     628           292           -53.50%
BenchmarkAllocProperty/props3     1033          443           -57.12%
BenchmarkAllocProperty/props4     1485          624           -57.98%
BenchmarkAllocProperty/props5     2012          828           -58.85%
BenchmarkAllocProperty/props6     2573          1062          -58.73%
BenchmarkAllocProperty/props7     3201          1375          -57.04%
BenchmarkAllocProperty/props8     3829          1591          -58.45%

benchmark                         old allocs     new allocs     delta
BenchmarkAllocProperty/props0     2              1              -50.00%
BenchmarkAllocProperty/props1     5              3              -40.00%
BenchmarkAllocProperty/props2     8              6              -25.00%
BenchmarkAllocProperty/props3     11             10             -9.09%
BenchmarkAllocProperty/props4     14             14             +0.00%
BenchmarkAllocProperty/props5     17             19             +11.76%
BenchmarkAllocProperty/props6     20             25             +25.00%
BenchmarkAllocProperty/props7     23             32             +39.13%
BenchmarkAllocProperty/props8     26             37             +42.31%

benchmark                         old bytes     new bytes     delta
BenchmarkAllocProperty/props0     176           96            -45.45%
BenchmarkAllocProperty/props1     672           240           -64.29%
BenchmarkAllocProperty/props2     1168          432           -63.01%
BenchmarkAllocProperty/props3     1664          672           -59.62%
BenchmarkAllocProperty/props4     2160          912           -57.78%
BenchmarkAllocProperty/props5     2656          1200          -54.82%
BenchmarkAllocProperty/props6     3152          1536          -51.27%
BenchmarkAllocProperty/props7     3648          1920          -47.37%
BenchmarkAllocProperty/props8     4144          2208          -46.72%
```