@funny-falcon-at-joomcode 2018-12-03
Added commit "Move "underlying errors" into property".
New comparison:
```
benchmark                         old ns/op     new ns/op     delta
BenchmarkAllocProperty/props0     100           67.7          -32.30%
BenchmarkAllocProperty/props1     319           148           -53.61%
BenchmarkAllocProperty/props2     628           262           -58.28%
BenchmarkAllocProperty/props3     1033          375           -63.70%
BenchmarkAllocProperty/props4     1485          530           -64.31%
BenchmarkAllocProperty/props5     2012          713           -64.56%
BenchmarkAllocProperty/props6     2573          969           -62.34%
BenchmarkAllocProperty/props7     3201          1099          -65.67%
BenchmarkAllocProperty/props8     3829          1297          -66.13%

benchmark                         old allocs     new allocs     delta
BenchmarkAllocProperty/props0     2              1              -50.00%
BenchmarkAllocProperty/props1     5              3              -40.00%
BenchmarkAllocProperty/props2     8              6              -25.00%
BenchmarkAllocProperty/props3     11             9              -18.18%
BenchmarkAllocProperty/props4     14             13             -7.14%
BenchmarkAllocProperty/props5     17             18             +5.88%
BenchmarkAllocProperty/props6     20             24             +20.00%
BenchmarkAllocProperty/props7     23             28             +21.74%
BenchmarkAllocProperty/props8     26             33             +26.92%

benchmark                         old bytes     new bytes     delta
BenchmarkAllocProperty/props0     176           64            -63.64%
BenchmarkAllocProperty/props1     672           176           -73.81%
BenchmarkAllocProperty/props2     1168          336           -71.23%
BenchmarkAllocProperty/props3     1664          496           -70.19%
BenchmarkAllocProperty/props4     2160          704           -67.41%
BenchmarkAllocProperty/props5     2656          960           -63.86%
BenchmarkAllocProperty/props6     3152          1264          -59.90%
BenchmarkAllocProperty/props7     3648          1472          -59.65%
BenchmarkAllocProperty/props8     4144          1728          -58.30%
```