@funny-falcon-at-joomcode 2018-12-03
And made properties implemented as linked list:
```
benchmark                         old ns/op     new ns/op     delta
BenchmarkAllocProperty/props0     100           68.2          -31.80%
BenchmarkAllocProperty/props1     319           135           -57.68%
BenchmarkAllocProperty/props2     628           202           -67.83%
BenchmarkAllocProperty/props3     1033          275           -73.38%
BenchmarkAllocProperty/props4     1485          341           -77.04%
BenchmarkAllocProperty/props5     2012          408           -79.72%
BenchmarkAllocProperty/props6     2573          474           -81.58%
BenchmarkAllocProperty/props7     3201          540           -83.13%
BenchmarkAllocProperty/props8     3829          607           -84.15%

benchmark                         old allocs     new allocs     delta
BenchmarkAllocProperty/props0     2              1              -50.00%
BenchmarkAllocProperty/props1     5              3              -40.00%
BenchmarkAllocProperty/props2     8              5              -37.50%
BenchmarkAllocProperty/props3     11             7              -36.36%
BenchmarkAllocProperty/props4     14             9              -35.71%
BenchmarkAllocProperty/props5     17             11             -35.29%
BenchmarkAllocProperty/props6     20             13             -35.00%
BenchmarkAllocProperty/props7     23             15             -34.78%
BenchmarkAllocProperty/props8     26             17             -34.62%

benchmark                         old bytes     new bytes     delta
BenchmarkAllocProperty/props0     176           64            -63.64%
BenchmarkAllocProperty/props1     672           160           -76.19%
BenchmarkAllocProperty/props2     1168          256           -78.08%
BenchmarkAllocProperty/props3     1664          352           -78.85%
BenchmarkAllocProperty/props4     2160          448           -79.26%
BenchmarkAllocProperty/props5     2656          544           -79.52%
BenchmarkAllocProperty/props6     3152          640           -79.70%
BenchmarkAllocProperty/props7     3648          736           -79.82%
BenchmarkAllocProperty/props8     4144          832           -79.92%
```